moving Activity.GetStatus method to DurableTask.ApplicationInsights pckg
