+++ src/DurableTask.ApplicationInsights/DurableTelemetryModule.cs
@@ -15,7 +15,6 @@ using System;
 using System.Diagnostics;
 using System.Threading;
 using System.Threading.Tasks;
-using DurableTask.Core.Tracing;
 using Microsoft.ApplicationInsights;
 using Microsoft.ApplicationInsights.DataContracts;
 using Microsoft.ApplicationInsights.Extensibility;
