+++ src/DurableTask.Core/Tracing/DiagnosticActivityExtensions.cs
@@ -22,29 +22,21 @@ namespace DurableTask.Core.Tracing
     /// <summary>
     /// Replica from System.Diagnostics.DiagnosticSource >= 6.0.0
     /// </summary>
-    public enum ActivityStatusCode
+    internal enum ActivityStatusCode
     {
-        ///<summary>Unset</summary>
         Unset = 0,
-        ///<summary>OK</summary>
         OK = 1,
-        ///<summary>Error</summary>
         Error = 2,
     }
 
     /// <summary>
     /// Extensions for <see cref="Activity"/>.
     /// </summary>
-    public static class DiagnosticActivityExtensions
+    internal static class DiagnosticActivityExtensions
     {
-        private static readonly string UnsetStatusCodeTagValue = "UNSET";
-        private static readonly string OkStatusCodeTagValue = "OK";
-        private static readonly string ErrorStatusCodeTagValue = "ERROR";
-
         private static readonly Action<Activity, string> s_setSpanId;
         private static readonly Action<Activity, string> s_setId;
         private static readonly Action<Activity, ActivityStatusCode, string> s_setStatus;
-        private static Func<Activity, ActivityStatusCode> s_getStatus;
 
         static DiagnosticActivityExtensions()
         {
@@ -52,40 +44,16 @@ namespace DurableTask.Core.Tracing
             s_setSpanId = typeof(Activity).GetField("_spanId", flags).CreateSetter<Activity, string>();
             s_setId = typeof(Activity).GetField("_id", flags).CreateSetter<Activity, string>();
             s_setStatus = CreateSetStatus();
-            s_getStatus = CreateGetStatus;
         }
 
-        /// <summary>
-        /// Set ID
-        /// </summary>
-        /// <param name="activity"></param>
-        /// <param name="id"></param>
         public static void SetId(this Activity activity, string id)
             => s_setId(activity, id);
 
-        /// <summary>
-        /// Set Span ID
-        /// </summary>
-        /// <param name="activity"></param>
-        /// <param name="spanId"></param>
         public static void SetSpanId(this Activity activity, string spanId)
             => s_setSpanId(activity, spanId);
 
-        /// <summary>
-        /// Set Status
-        /// </summary>
-        /// <param name="activity"></param>
-        /// <param name="status"></param>
-        /// <param name="description"></param>
         public static void SetStatus(this Activity activity, ActivityStatusCode status, string description)
             => s_setStatus(activity, status, description);
-
-        /// <summary>
-        /// Get Status
-        /// </summary>
-        /// <param name="activity"></param>
-        public static ActivityStatusCode GetStatus(this Activity activity)
-            => s_getStatus(activity);
 
         private static Action<Activity, ActivityStatusCode, string> CreateSetStatus()
         {
@@ -116,93 +84,6 @@ namespace DurableTask.Core.Tracing
             MethodCallExpression callExp = Expression.Call(targetExp, method, convert, description);
             return Expression.Lambda<Action<Activity, ActivityStatusCode, string>>(callExp, targetExp, status, description)
                 .Compile();
-        }
-
-        /// <summary>
-        /// Create Get Status
-        /// </summary>
-        /// <param name="activity"></param>
-        /// <returns></returns>
-        public static ActivityStatusCode CreateGetStatus(Activity activity)
-        {
-            return !activity.TryGetStatus(out ActivityStatusCode statusCode, out string statusDescription)
-                    ? ActivityStatusCode.Unset
-                    : statusCode;
-        }
-
-        /// <summary>
-        /// Try Get Status
-        /// </summary>
-        /// <param name="activity"></param>
-        /// <param name="statusCode"></param>
-        /// <param name="statusDescription"></param>
-        /// <returns></returns>
-        public static bool TryGetStatus(this Activity activity, out ActivityStatusCode statusCode, out string statusDescription)
-        {
-            Debug.Assert(activity != null, "Activity should not be null");
-
-            bool foundStatusCode = false;
-            statusCode = default;
-            statusDescription = null;
-
-            foreach (var tag in activity.Tags)
-            {
-                switch (tag.Key)
-                {
-                    case Schema.Status.Code:
-                        foundStatusCode = TryGetStatusCodeForTagValue(tag.Value as string, out statusCode);
-                        if (!foundStatusCode)
-                        {
-                            // If status code was found but turned out to be invalid give up immediately.
-                            return false;
-                        }
-
-                        break;
-                    case Schema.Status.Description:
-                        statusDescription = tag.Value as string;
-                        break;
-                    default:
-                        continue;
-                }
-
-                if (foundStatusCode && statusDescription != null)
-                {
-                    // If we found a status code and a description we break enumeration because our work is done.
-                    break;
-                }
-            }
-
-            return foundStatusCode;
-        }
-
-        /// <summary>
-        /// Try Get Status Code
-        /// </summary>
-        /// <param name="statusCodeTagValue"></param>
-        /// <param name="statusCode"></param>
-        /// <returns></returns>
-        public static bool TryGetStatusCodeForTagValue(string statusCodeTagValue, out ActivityStatusCode statusCode)
-        {
-            ActivityStatusCode? tempStatusCode = GetStatusCodeForTagValue(statusCodeTagValue);
-
-            statusCode = tempStatusCode ?? default;
-
-            return tempStatusCode.HasValue;
-        }
-        private static ActivityStatusCode? GetStatusCodeForTagValue(string statusCodeTagValue)
-        {
-            return statusCodeTagValue switch
-            {
-                /*
-                 * Note: Order here does matter for perf. Unset is
-                 * first because assumption is most spans will be
-                 * Unset, then Error. Ok is not set by the SDK.
-                 */
-                string _ when UnsetStatusCodeTagValue.Equals(statusCodeTagValue, StringComparison.OrdinalIgnoreCase) => ActivityStatusCode.Unset,
-                string _ when ErrorStatusCodeTagValue.Equals(statusCodeTagValue, StringComparison.OrdinalIgnoreCase) => ActivityStatusCode.Error,
-                string _ when OkStatusCodeTagValue.Equals(statusCodeTagValue, StringComparison.OrdinalIgnoreCase) => ActivityStatusCode.OK,
-                _ => null,
-            };
         }
     }
 }
